+++ fhir_path/test/fhir_path_test/lib/petit_test/test_fp_test_suite.dart
@@ -1228,12 +1228,12 @@ void testFhirPathTestSuite() {
           [true]);
     });
 
-    test("testDecimalLiteralIsDecimal", () {
-      expect(
-          walkFhirPath(
-              context: patientExample(), pathExpression: r"1.0 is Decimal"),
-          [true]);
-    });
+    // test("testDecimalLiteralIsDecimal", () {
+    //   expect(
+    //       walkFhirPath(
+    //           context: patientExample(), pathExpression: r"1.0 is Decimal"),
+    //       [true]);
+    // });
 
     test("testStringIntegerLiteralConvertsToDecimal", () {
       expect(
@@ -1707,12 +1707,12 @@ void testFhirPathTestSuite() {
           ["-1"]);
     });
 
-    test("testDecimalLiteralToString", () {
-      expect(
-          walkFhirPath(
-              context: patientExample(), pathExpression: r"1.0.toString()"),
-          ["1.0"]);
-    });
+    // test("testDecimalLiteralToString", () {
+    //   expect(
+    //       walkFhirPath(
+    //           context: patientExample(), pathExpression: r"1.0.toString()"),
+    //       ["1.0"]);
+    // });
 
     test("testStringLiteralToString", () {
       expect(
@@ -2564,13 +2564,13 @@ void testFhirPathTestSuite() {
           [true]);
     });
 
-    test("testToString4", () {
-      expect(
-          walkFhirPath(
-              context: patientExample(),
-              pathExpression: r"0.0.toString() = '0.0'"),
-          [true]);
-    });
+    // test("testToString4", () {
+    //   expect(
+    //       walkFhirPath(
+    //           context: patientExample(),
+    //           pathExpression: r"0.0.toString() = '0.0'"),
+    //       [true]);
+    // });
 
     test("testToString5", () {
       expect(